Revert "Split neo4j mapper into two files."
This reverts commit 2643348aed98d3ac22485a553873035bd84ffcd2.
