+++ stellargraph/layer/graphsage.py
@@ -47,11 +47,9 @@ from ..mapper import (
 )
 
 from .misc import deprecated_model_function
-from ..connector.neo4j.mapper_nodes import (
+from ..connector.neo4j.mapper import (
     Neo4jGraphSAGENodeGenerator,
     Neo4jDirectedGraphSAGENodeGenerator,
-)
-from ..connector.neo4j.mapper_links import (
     Neo4jGraphSAGELinkGenerator,
     Neo4jDirectedGraphSAGELinkGenerator,
 )
